+++ app/ante/eth_test.go
@@ -195,7 +195,7 @@ func (suite AnteTestSuite) TestEthNonceVerificationDecorator() {
 }
 
 func (suite AnteTestSuite) TestEthGasConsumeDecorator() {
-	dec := ante.NewEthGasConsumeDecorator(suite.app.EvmKeeper, suite.app.FeeMarketKeeper)
+	dec := ante.NewEthGasConsumeDecorator(suite.app.EvmKeeper)
 
 	addr := tests.GenerateAddress()